backend/eunoia_backend/requirements.txt
Django
python-dotenv
openai
openai-agents
Pillow
djangorestframework 
django-cors-headers
